requirements.txt
Django==1.9.7
djangorestframework==3.4.0
gunicorn==19.9.0
dj-database-url==0.3.0
dj-static==0.0.6
Unipath==1.0
python-decouple==3
Pillow==3.3.0
Markdown==2.6.6
bleach==1.4.3
psycopg2==2.7.5
whitenoise==3.3.1
